refactor(db): use importFile instead of query when possible.
Rather than treating everything as a query, prefer to use the
`importFile` method, which is similar to calling `SOURCE` on the
`mariadb` CLI. Hence it simplifies and makes it more clear that what is
going on is the import of valid `.sql` scripts as a whole.

The reason why the `./database/Schema/Scheme.sql` and
`.database/Schema/Triggers.sql` files are still read as files and
treated as `multipleStatement` queries is because I found no way to
implement them with `importFile`.

The method seems to lack the ability to deal with complex trigger
declarations which involve `BEGIN` and `END`, and multiple statements.
And when it comes to the creation of the database, I'm unsure as to why
it fails. But it seems to work with the already implemented
`fs.readFileSync` method.