docs(readme): new db setup instructions
The introductiion of these scripts makes it so that the `.sql` scripts
are no longer necessarily CLI-compatible. Specifically the
`database/Schema/Triggers.sql` file. But the whole point of this branch
is to make it so all the necessary scripts are run automatically. Hence
it's best to have that be the outlined way to set up the database.

That required reordering the sections such that the environment
variables are setup first, since they are needed for the setup scripts.
And removing the old batch way of loading the `.sql` files for this
newer and sleaker version.