chore(pr_templates): eliminate individual pr templates
The JS Copy-Paste detector (jscpd) linter didn't appreciate the
inclusion of these files, given that it is identical to the contents of
the generic `pull_request_template.md` file.

It does get rid of the ability for `gh` CLI users to be able to directly
query a specific template, but definitely sounds like a niche use case
that is not as important.